slamd-0 - mov tests to root and add test for liquids
